+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "files.associations": {
+        "stdio.h": "c",
+        "menu_principal.h": "c"
+    }
+}
