+++ menu_principal.c
@@ -1,0 +1,29 @@
+char menu_principal(void) {
+    system("cls");
+
+    printf("\t\t=====================================\n");
+    printf("\t\t||           SIG-Customer          ||\n");
+    printf("\t\t=====================================\n");
+    printf("\t\t-------------------------------------\n");
+    printf("\t\t|       1 - Modulo de Cliente       |\n");
+    printf("\t\t-------------------------------------\n");
+    printf("\t\t-------------------------------------\n");
+    printf("\t\t|   2 - Modulo em Desenvolvimento   |\n");
+    printf("\t\t-------------------------------------\n");
+    printf("\t\t-------------------------------------\n");
+    printf("\t\t|   3 - Modulo em Desenvolvimento   |\n");
+    printf("\t\t-------------------------------------\n");
+    printf("\t\t-------------------------------------\n");
+    printf("\t\t|        4 - Sobre e Equipe         |\n");
+    printf("\t\t-------------------------------------\n");
+    printf("\t\t-------------------------------------\n");
+    printf("\t\t|           0 - Encerrar            |\n");
+    printf("\t\t-------------------------------------\n");
+
+    char op;
+
+    printf("\nDigite a opcao desejada: ");
+    scanf("%s", &op);
+
+    return op;
+}
