+++ sobre.c
@@ -1,0 +1,19 @@
+void menu_sobre(void) {
+    system("cls");
+
+    printf("\n\n");
+    printf("\t\tSobre o projeto:\n");
+    printf("\n");
+    printf("\tSIG-Custumer: Um sistema de controle de cliente\n");
+    printf("As empresas da atualidade utilizam complexos sistemas gerenciais para,\n");
+    printf("entre varias outras coisas, controlar seus clientes\n");
+    printf("alem de ajudar a conseguir novos clientes.\n");
+    printf("O presente projeto tem como objetivo propor um software que implementa\n");
+    printf("funcionalidades basicas relacionadas ao controle de clientes de uma empresa,\n");
+    printf("facilitando a gestao\n");
+    printf("\n\n");
+    printf("\t\tDiscentes/Autores do Projeto:\n\n");
+    printf("\tCharles Eduardo Araujo De Faria. Matricula: 20220042687\n");
+    printf("\tVictor Ryan Galvao Silva. Matricula: 20220035163\n");
+    printf("\n");
+}
